+++ Simple Game Library/src/at/fhooe/mtd/sgl/input/Util.java
@@ -1,0 +1,13 @@
+package at.fhooe.mtd.sgl.input;
+
+import java.awt.event.KeyEvent;
+
+public class Util {
+    
+    public static boolean isPrintableChar(char c) {
+        Character.UnicodeBlock block = Character.UnicodeBlock.of(c);
+        return (!Character.isISOControl(c)) && c != KeyEvent.CHAR_UNDEFINED
+                && block != null && block != Character.UnicodeBlock.SPECIALS;
+    }
+    
+}